requirements.txt
streamlit
requests
pandas
google-cloud-bigquery
pandas-gbq
pytest
db-dtypes
google-genai
google-cloud-aiplatform
google-vertexai
google-genai
